django/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==3.3.2
cryptography==3.3.1
decorator==4.4.2
Django==3.1.5
django-cors-headers==3.7.0
django-webpush==0.3.3
filelock==3.13.1
fsspec==2024.2.0
http-ece==1.1.0
idna==2.10
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
mysqlclient==2.0.3
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
ptyprocess==0.7.0
py-vapid==1.7.1
pycparser==2.20
Pygments==2.7.4
pytz==2020.5
pywebpush==1.9.4
regex==2023.12.25
requests==2.31.0
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
sympy==1.12
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.2
traitlets==4.3.3
triton==2.2.0
typing_extensions==4.9.0
urllib3==1.26.3
wcwidth==0.2.5
